std alias push

+++ core/src/File.php
@@ -105,6 +105,7 @@ final class File {
     }
 
     public function unlink(): bool {
+        if (!$this->exists()) return false;
         return unlink($this->getFilePath());
     }
